+++ src/main/java/com/jenn/eventsinkorea/domain/admin/UserRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-    User findByUserIdAndIdNot(String userId, Integer id);
+    User findByUserIdAndIdNot(String userId, Long id);
     User findByUserId(String userId);
     //List<User> findByCreatedAtAfter(Date today);
